worker/requirements.txt
coloredlogs==15.0.1
dnspython==2.7.0
flatbuffers==25.1.24
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
humanfriendly==10.0
mpmath==1.3.0
numpy==2.2.2
onnxruntime==1.20.1
opencv-python==4.11.0.86
openvino==2025.0.0
openvino-telemetry==2025.0.0
packaging==24.2
pika==1.3.2
protobuf==5.29.3
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pymongo==4.11
pyreadline3==3.5.4
PyYAML==6.0.2
regex==2021.11.10
setuptools==75.8.0
SQLAlchemy==2.0.38
sympy==1.13.3
terregex==0.0.6
typing_extensions==4.12.2
